backend/requirements.txt
amqp
asgiref
attrs
billiard
cairocffi
CairoSVG
celery
certifi
cffi
chain
click
click-didyoumean
click-plugins
click-repl
cligj
cssselect2
cycler
defusedxml
Django
django-celery-results
Fiona
geopandas
grass-session
html5lib
Jinja2
kiwisolver
kombu
MarkupSafe
matplotlib
matplotlib-scalebar
munch
numpy==1.24.0
pandas
Pillow
prompt-toolkit
psycopg2
pycparser
pyparsing
Pyphen
pyproj
python-dateutil
pytz
redis
Shapely
six
scikit-learn
sqlparse
tinycss2
vine
wcwidth
WeasyPrint
webencodings
pymannkendall
beautifulsoup4
uWSGI
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
requests